# backend/openapi.json
{
 "components": {
  "schemas": {
...
   "CaptureConfig-Input": {
    "properties": {
     "categories": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Categories"
     },
     "device_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Id"
     },
     "include_telemetry": {
      "default": true,
      "title": "Include Telemetry",
      "type": "boolean"
     },
     "max_bytes": {
      "default": 33554432,
      "maximum": 268435456.0,
      "minimum": 65536.0,
      "title": "Max Bytes",
      "type": "integer"
     },
     "max_events": {
      "default": 50000,
      "maximum": 500000.0,
      "minimum": 100.0,
      "title": "Max Events",
      "type": "integer"
     },
     "max_seconds": {
      "default": 900,
      "maximum": 86400.0,
      "minimum": 5.0,
      "title": "Max Seconds",
      "type": "integer"
     },
     "min_severity": {
      "anyOf": [
       {
        "enum": [
         "debug",
         "info",
         "notice",
         "warning",
         "error",
         "critical"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Min Severity"
     },
     "name": {
      "default": "capture",
      "maxLength": 80,
      "title": "Name",
      "type": "string"
     },
     "sources": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sources"
     },
     "trace_events": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "description": "start a tracefs session for the capture",
      "title": "Trace Events"
     }
    },
    "title": "CaptureConfig",
    "type": "object"
   },
...
   "CaptureInfo": {
    "properties": {
     "artifacts": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Artifacts",
      "type": "array"
     },
     "config": {
      "$ref": "#/components/schemas/CaptureConfig-Output"
     },
     "created_at": {
      "title": "Created At",
      "type": "number"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "mode": {
      "title": "Mode",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "seconds_remaining": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seconds Remaining"
     },
     "started_ns": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started Ns"
     },
     "state": {
      "enum": [
       "running",
       "stopped",
       "expired",
       "error"
      ],
      "title": "State",
      "type": "string"
     },
     "stats": {
      "$ref": "#/components/schemas/CaptureStats"
     },
     "stopped_ns": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Stopped Ns"
     }
    },
    "required": [
     "id",
     "name",
     "state",
     "created_at",
     "started_ns",
     "stopped_ns",
     "mode",
     "config",
     "stats",
     "artifacts",
     "seconds_remaining"
    ],
    "title": "CaptureInfo",
    "type": "object"
   },
...
   "Device": {
    "properties": {
     "bus": {
      "enum": [
       "pci",
       "usb",
       "sdio",
       "platform",
       "other"
      ],
      "title": "Bus",
      "type": "string"
     },
     "bus_path": {
      "items": {
       "$ref": "#/components/schemas/BusNode"
      },
      "title": "Bus Path",
      "type": "array"
     },
     "demo": {
      "default": false,
      "title": "Demo",
      "type": "boolean"
     },
     "detection": {
      "items": {
       "$ref": "#/components/schemas/DetectionReason"
      },
      "title": "Detection",
      "type": "array"
     },
     "driver": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Driver"
     },
     "driver_info": {
      "$ref": "#/components/schemas/DriverInfo"
     },
     "extensions": {
      "additionalProperties": true,
      "title": "Extensions",
      "type": "object"
     },
     "firmware": {
      "$ref": "#/components/schemas/FirmwareInfo"
     },
     "firmware_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Firmware Version"
     },
     "first_seen": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "First Seen"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "last_seen": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Seen"
     },
     "module": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Module"
     },
     "netdev_info": {
      "items": {
       "$ref": "#/components/schemas/Netdev"
      },
      "title": "Netdev Info",
      "type": "array"
     },
     "netdevs": {
      "items": {
       "type": "string"
      },
      "title": "Netdevs",
      "type": "array"
     },
     "operstate": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Operstate"
     },
     "pci": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/PciInfo"
       },
       {
        "type": "null"
       }
      ]
     },
     "pci_config": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/PciConfig"
       },
       {
        "type": "null"
       }
      ]
     },
     "phys": {
      "items": {
       "type": "string"
      },
      "title": "Phys",
      "type": "array"
     },
     "power": {
      "$ref": "#/components/schemas/PowerInfo"
     },
     "present": {
      "default": true,
      "title": "Present",
      "type": "boolean"
     },
     "product_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Id"
     },
     "product_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Name"
     },
     "rfkill": {
      "items": {
       "$ref": "#/components/schemas/Rfkill"
      },
      "title": "Rfkill",
      "type": "array"
     },
     "sysfs_path": {
      "title": "Sysfs Path",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "usb": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/UsbInfo"
       },
       {
        "type": "null"
       }
      ]
     },
     "vendor_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vendor Id"
     },
     "vendor_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vendor Name"
     },
     "wiphys": {
      "items": {
       "$ref": "#/components/schemas/Wiphy"
      },
      "title": "Wiphys",
      "type": "array"
     }
    },
    "required": [
     "id",
     "bus",
     "sysfs_path",
     "title",
     "vendor_id",
     "product_id",
     "vendor_name",
     "product_name",
     "driver",
     "module",
     "firmware_version",
     "phys",
     "netdevs",
     "operstate",
     "detection",
     "first_seen",
     "last_seen",
     "present",
     "demo",
     "bus_path",
     "pci",
     "pci_config",
     "usb",
     "driver_info",
     "firmware",
     "wiphys",
     "netdev_info",
     "power",
     "rfkill",
     "extensions"
    ],
    "title": "Device",
    "type": "object"
   },
...
   "Event": {
    "properties": {
     "category": {
      "enum": [
       "kernel_log",
       "pci_error",
       "pci_link",
       "usb_error",
       "usb_transfer",
       "hotplug",
       "driver_state",
       "firmware",
       "association",
       "phy",
       "mlo",
       "reset",
       "power",
       "scan",
       "regulatory",
       "trace",
       "telemetry",
       "netdev",
       "diagnostic",
       "system"
      ],
      "title": "Category",
      "type": "string"
     },
     "data": {
      "additionalProperties": true,
      "title": "Data",
      "type": "object"
     },
     "demo": {
      "default": false,
      "title": "Demo",
      "type": "boolean"
     },
     "device_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Id"
     },
     "explanation": {
      "title": "Explanation",
      "type": "string"
     },
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "raw": {
      "title": "Raw",
      "type": "string"
     },
     "session": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session"
     },
     "severity": {
      "default": "info",
      "enum": [
       "debug",
       "info",
       "notice",
       "warning",
       "error",
       "critical"
      ],
      "title": "Severity",
      "type": "string"
     },
     "source": {
      "title": "Source",
      "type": "string"
     },
     "summary": {
      "title": "Summary",
      "type": "string"
     },
     "ts_boottime_ns": {
      "description": "Canonical timestamp: CLOCK_BOOTTIME nanoseconds",
      "title": "Ts Boottime Ns",
      "type": "integer"
     },
     "ts_raw": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Original timestamp in its own clock domain",
      "title": "Ts Raw"
     },
     "ts_source": {
      "enum": [
       "tracefs_boot",
       "tracefs_mono",
       "kernel_printk",
       "journal_monotonic",
       "receive_boottime",
       "poll_boottime",
       "netlink_boottime",
       "usbmon_monotonic",
       "recorded"
      ],
      "title": "Ts Source",
      "type": "string"
     },
     "ts_wall": {
      "description": "Derived wall clock (unix seconds) - display only",
      "title": "Ts Wall",
      "type": "number"
     }
    },
    "required": [
     "id",
     "ts_boottime_ns",
     "ts_source",
     "ts_raw",
     "ts_wall",
     "source",
     "device_id",
     "severity",
     "category",
     "kind",
     "summary",
     "explanation",
     "raw",
     "data",
     "demo",
     "session"
    ],
    "title": "Event",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "LoginBody": {
    "properties": {
     "token": {
      "title": "Token",
      "type": "string"
     }
    },
    "required": [
     "token"
    ],
    "title": "LoginBody",
    "type": "object"
   },
...
   "ReplayBody": {
    "properties": {
     "speed": {
      "default": 1.0,
      "maximum": 20.0,
      "minimum": 0.1,
      "title": "Speed",
      "type": "number"
     }
    },
    "title": "ReplayBody",
    "type": "object"
   },
...
   "TelemetrySample": {
    "description": "A periodic numeric sample (for charts). Not an event; kept in a bounded ring.",
    "properties": {
     "demo": {
      "default": false,
      "title": "Demo",
      "type": "boolean"
     },
     "device_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Id"
     },
     "series": {
      "title": "Series",
      "type": "string"
     },
     "ts_boottime_ns": {
      "title": "Ts Boottime Ns",
      "type": "integer"
     },
     "ts_wall": {
      "title": "Ts Wall",
      "type": "number"
     },
     "values": {
      "additionalProperties": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ]
      },
      "title": "Values",
      "type": "object"
     }
    },
    "required": [
     "ts_boottime_ns",
     "ts_wall",
     "device_id",
     "series",
     "values",
     "demo"
    ],
    "title": "TelemetrySample",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Wireless Hardware Debugger",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/v1/auth/login": {
   "post": {
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "boolean"
         },
         "title": "Response Login Api V1 Auth Login Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Login"
   }
  },
...
  "/api/v1/captures": {
   "get": {
    "operationId": "list_captures_api_v1_captures_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/CaptureInfo"
         },
         "title": "Response List Captures Api V1 Captures Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Captures"
   },
   "post": {
    "operationId": "create_capture_api_v1_captures_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CaptureConfig-Input"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CaptureInfo"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Capture"
   }
  },
  "/api/v1/captures/{cid}": {
   "delete": {
    "operationId": "delete_capture_api_v1_captures__cid__delete",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "title": "Cid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "boolean"
         },
         "title": "Response Delete Capture Api V1 Captures  Cid  Delete",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Capture"
   },
   "get": {
    "operationId": "get_capture_api_v1_captures__cid__get",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "title": "Cid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CaptureInfo"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Capture"
   }
  },
  "/api/v1/captures/{cid}/bundle": {
   "get": {
    "operationId": "bundle_api_v1_captures__cid__bundle_get",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "title": "Cid",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "redact_macs",
      "required": false,
      "schema": {
       "default": true,
       "title": "Redact Macs",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Bundle"
   }
  },
  "/api/v1/captures/{cid}/events": {
   "get": {
    "operationId": "capture_events_api_v1_captures__cid__events_get",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "title": "Cid",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "title": "Offset",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 2000,
       "maximum": 20000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Event"
         },
         "title": "Response Capture Events Api V1 Captures  Cid  Events Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Capture Events"
   }
  },
  "/api/v1/captures/{cid}/export": {
   "get": {
    "operationId": "export_capture_api_v1_captures__cid__export_get",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "title": "Cid",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "format",
      "required": false,
      "schema": {
       "default": "json",
       "pattern": "^(json|jsonl|csv|trace)$",
       "title": "Format",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Export Capture"
   }
  },
  "/api/v1/captures/{cid}/replay": {
   "post": {
    "operationId": "start_replay_api_v1_captures__cid__replay_post",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "title": "Cid",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReplayBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Start Replay Api V1 Captures  Cid  Replay Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Start Replay"
   }
  },
  "/api/v1/captures/{cid}/stop": {
   "post": {
    "operationId": "stop_capture_api_v1_captures__cid__stop_post",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "title": "Cid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CaptureInfo"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Stop Capture"
   }
  },
  "/api/v1/captures/{cid}/telemetry": {
   "get": {
    "operationId": "capture_telemetry_api_v1_captures__cid__telemetry_get",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "title": "Cid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TelemetrySample"
         },
         "title": "Response Capture Telemetry Api V1 Captures  Cid  Telemetry Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Capture Telemetry"
   }
  },
...
  "/api/v1/devices/{device_id}": {
   "get": {
    "operationId": "get_device_api_v1_devices__device_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "device_id",
      "required": true,
      "schema": {
       "title": "Device Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Device"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Device"
   }
  },
  "/api/v1/devices/{device_id}/evidence": {
   "get": {
    "operationId": "device_evidence_api_v1_devices__device_id__evidence_get",
    "parameters": [
     {
      "in": "path",
      "name": "device_id",
      "required": true,
      "schema": {
       "title": "Device Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Device Evidence Api V1 Devices  Device Id  Evidence Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Device Evidence"
   }
  },
...
 }
}